chore: rename ring_buffer.ts to ring-buffer.ts
1. Inconsistent with test file ring-buffer.test.ts (hyphen vs underscore).
2. Rename source file and update two import paths accordingly.
