+++ src/ring-buffer.ts
